requirements.txt
google-adk
streamlit
sqlalchemy
google-genai
aiosqlite
greenlet